+++ app/views/layouts/application.html.erb
@@ -9,7 +9,6 @@
   <link rel="stylesheet" type="text/css" href="http://yui.yahooapis.com/3.18.1/build/cssreset/cssreset-min.css">
   <%= stylesheet_link_tag 'application', media: 'all'%>
   <%= javascript_pack_tag 'application' %>
-  <script src="index.js"></script>
 </head>
 
 <body>
